+++ R/utils.R
@@ -18,7 +18,7 @@ get_ipc <- function(folder = tempdir()){
   u <- "http://www.ine.gub.uy/c/document_library/get_file?uuid=2e92084a-94ec-4fec-b5ca-42b40d5d2826&groupId=10181"
   f <- fs::path(folder, "IPC gral var M_B10.xls")
   if (identical(.Platform$OS.type, "unix")) {
-    try(utils::download.file(u, f, mode = 'wb', method = 'wget', extra = '--no-check-certificate'))
+    try(utils::download.file(u, f, mode = 'wb', method = 'wget'))
   } else {
     try(utils::download.file(u, f, mode = 'wb', method = 'libcurl'))
   }
@@ -92,7 +92,7 @@ get_ipc_region <- function(folder = tempdir(), region, sheet = 1){
     f <- fs::path(folder, "IPC 3.2 indvarinc_ div M_B10_Int.xls")
   }
   if (identical(.Platform$OS.type, "unix")) {
-    try(utils::download.file(u, f, mode = 'wb', method = 'wget', extra = '--no-check-certificate'))
+    try(utils::download.file(u, f, mode = 'wb', method = 'wget'))
   } else {
     try(utils::download.file(u, f, mode = 'wb', method = 'libcurl'))
   }
@@ -144,7 +144,7 @@ get_cba_cbna <- function(folder = tempdir(), region, sheet = 1){
   u <- "http://www.ine.gub.uy/c/document_library/get_file?uuid=1675e7d0-6fe0-49bd-bf3f-a46bd6334c0c&groupId=10181"
   f <- fs::path(folder, "CBA_LP_LI_M.xls")
   if (identical(.Platform$OS.type, "unix")) {
-    try(utils::download.file(u, f, mode = 'wb', method = 'wget', extra = '--no-check-certificate'))
+    try(utils::download.file(u, f, mode = 'wb', method = 'wget'))
   } else {
     try(utils::download.file(u, f, mode = 'wb', method = 'libcurl'))
   }
@@ -208,7 +208,7 @@ get_ipab <- function(folder = tempdir(), sheet = 1){
   u <- "http://www.ine.gub.uy/c/document_library/get_file?uuid=c4b5efaa-cdd4-497a-ab78-e3138e4f08dc&groupId=10181"
   f <- fs::path(folder, "IPC Div M_B10.xls")
   if (identical(.Platform$OS.type, "unix")) {
-    try(utils::download.file(u, f, mode = 'wb', method = 'wget', extra = '--no-check-certificate'))
+    try(utils::download.file(u, f, mode = 'wb', method = 'wget'))
   } else {
     try(utils::download.file(u, f, mode = 'wb', method = 'libcurl'))
   }
@@ -252,7 +252,7 @@ get_ipab_region <- function(folder = tempdir(), region, sheet = 1){
     f <- fs::path(folder, "IPC 3.2 indvarinc_ div M_B10_Int.xls")
   }
   if (identical(.Platform$OS.type, "unix")) {
-    try(utils::download.file(u, f, mode = 'wb', method = 'wget', extra = '--no-check-certificate'))
+    try(utils::download.file(u, f, mode = 'wb', method = 'wget'))
   } else {
     try(utils::download.file(u, f, mode = 'wb', method = 'libcurl'))
   }
@@ -468,7 +468,7 @@ organize_ht11 <- function(data, year, ht11 = "ht11", numero = "numero") {
 #   u <- "http://www.ine.gub.uy/documents/10181/33330/CORRESPONDENCIA+CIUU4+A+CIUU3.pdf/623c43cb-009c-4da9-b48b-45282745063b"
 #   f <- fs::path(folder, "ciiu4.pdf")
 #   if (identical(.Platform$OS.type, "unix")) {
-#     try(utils::download.file(u, f, mode = 'wb', method = 'wget', extra = '--no-check-certificate'))
+#     try(utils::download.file(u, f, mode = 'wb', method = 'wget'))
 #   } else {
 #     try(utils::download.file(u, f, mode = 'wb', method = 'libcurl'))
 #   }